[ML] Extract job selection resolver (#95394)
